Allow adding labels to PRs
Unlike other lists, we can't split on `,` for the labels, as they can contain
anything on GitHub, even emojis :puke: Thus, we need to split on a newline, the
only character that a label can't contain. This, in turn, makes the code
unbelieveably simpler, since bash assumes everything to be split on a newline
:tada:

In order to properly send the data to GitHub's API, we need to rely on `jo`, as
manually constructing JSON payloads is for optimists.

Closes #5