Emit correct links to issues and PRs in the changelog
GitHub used #123 as a shorthand for a URL to the issue **in the same
repository**. Thus, when constructing the changelogs, all #123 in the changelog
would render as a link to the repository where niv-updater-action operates,
instead of pointing to the depdendency's repository.

To fix this issue, prepend anything that looks like #<numbers> with the name of
the dependency (owner/repository).

Closes #8.