All errors are fatal
I was mistaken in thinking there should be a distinction between an error and a
fatal event. The thinking was to let the script run even in the presence of
errors. However, this is wrong. If there is an error, the run should be marked
as a failure, and people should be notified, because something is definitely amis.