Fix generated commit message
git commit messages start with the subject, followed by an empty line and the body
of the commit. Due to old refactoring, the generated commit messages lost the
subject line, as that was now in the title, that is a separate file, used to set
the title of the PR. This change stores the commit message in a separate file
and generates it following the convention.