Add option to avoid GitHub backreferencing the PR
Closes #26

niv-update-action changelog in a GitHub PR contains references (in the form
of `owner/repo#PR_number`), which GitHub nicely turns into links.
However, what GitHub also does is it creates a backreference in those
PRs back to the PR created by niv-update-action. This pollutes the
upstream projects. Thus, this change adds an option to use a redirector
service in order to still have links, but avoid having backreferences.