Reformat mentions to reduce spamming of mentioned developers
If the changelog contains anyones username (@ + username), GitHub will dutifully
notify that person that they have been mentioned. This is not nice for our
generated changelogs, as it spams people and brings no value (the mention is not
intentional). To retain the display, yet avoid notification, we print '@'
followed by an UNICODE WORD JOINER character (that has zero width, but prevents
splitting at this boundary), followed by the username.