Better handling of GitHub SSH links
As these links come in many forms:

- `git+ssh://git@github.com/foo/bar`
- `ssh://git@github.com:foo/bar.git`
- `git@github.com:foo/bar.git`

This change takes care of properly handling each of them, so we could cover all
possible instances.

The code assumes that the content of `sources.json` is correct, and that `niv`
already verified that links are good. Thus, the checks and extractions are more
lenient than they should be.